+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - Implement Pushover emergency alert cancellation when check goes up
+- Add "The following checks are also down" section in Telegram notifications
 
 ### Bug Fixes
 - Fix report templates to not show the "started" status (show UP or DOWN instead)